Move sourceInfoHeading out of Splash
Components in this file are not used on Splash and should be directly
under src/components.
