+++ static-site/src/sections/community-repo-page.jsx
@@ -6,7 +6,7 @@ import DatasetSelect from "../components/Datasets/dataset-select";
 import GenericPage from "../layouts/generic-page";
 import { fetchAndParseJSON } from "../util/datasetsHelpers";
 import { ErrorBanner } from "../components/splash/errorMessages";
-import SourceInfoHeading from "../components/splash/sourceInfoHeading";
+import SourceInfoHeading from "../components/sourceInfoHeading";
 
 class Index extends React.Component {
   constructor(props) {
