+++ static-site/src/sections/individual-group-page.jsx
@@ -6,7 +6,7 @@ import * as splashStyles from "../components/splash/styles";
 import DatasetSelect from "../components/Datasets/dataset-select";
 import GenericPage from "../layouts/generic-page";
 import { fetchAndParseJSON } from "../util/datasetsHelpers";
-import SourceInfoHeading from "../components/splash/sourceInfoHeading";
+import SourceInfoHeading from "../components/sourceInfoHeading";
 import { ErrorBanner } from "../components/splash/errorMessages";
 import { canUserEditGroupSettings } from "./group-settings-page";
